refactor(test): move slow vector index tests to integration tests
Move compute-heavy vector index tests from unit test modules to
integration tests where they run with opt-level=1 for better
performance. This prevents CI timeouts caused by debug build slowness.

Tests migrated:
- test_build_ivf_pq, test_build_ivf_pq_v3, test_build_ivf_pq_4bit
- test_create_ivf_hnsw_pq, test_create_ivf_hnsw_pq_4bit
- test_optimize_with_empty_partition, test_remap_join_on_second_delta
- test_create_index_with_many_invalid_vectors
- test_build_pq_model_l2, test_build_pq_model_cosine
