+++ rust/lance/src/index/vector/ivf/v2.rs
@@ -665,7 +665,7 @@ mod tests {
     };
     use lance_linalg::distance::{multivec_distance, DistanceType};
     use lance_linalg::kernels::normalize_fsl;
-    use lance_testing::datagen::{generate_random_array, generate_random_array_with_range};
+    use lance_testing::datagen::generate_random_array_with_range;
     use object_store::path::Path;
     use rand::distr::uniform::SampleUniform;
     use rand::{rngs::StdRng, Rng, SeedableRng};
@@ -787,39 +787,6 @@ mod tests {
         }
         let schema: Arc<_> = Schema::new(fields).into();
         let batch = RecordBatch::try_new(schema.clone(), arrays).unwrap();
-        (batch, schema)
-    }
-
-    fn generate_clustered_batch(
-        rows_per_partition: usize,
-        offsets: [f32; 2],
-    ) -> (RecordBatch, SchemaRef) {
-        let num_partitions = offsets.len();
-        let total_rows = rows_per_partition * num_partitions;
-        let mut ids = Vec::with_capacity(total_rows);
-        let mut values = Vec::with_capacity(total_rows * DIM);
-        let mut rng = StdRng::seed_from_u64(42);
-        for (cluster_idx, offset) in offsets.iter().enumerate() {
-            for row in 0..rows_per_partition {
-                ids.push((cluster_idx * rows_per_partition + row) as u64);
-                for dim in 0..DIM {
-                    let base = if dim == 0 { *offset } else { 0.0 };
-                    let noise = (rng.random::<f32>() - 0.5) * 0.02;
-                    values.push(base + noise);
-                }
-            }
-        }
-        let ids = Arc::new(UInt64Array::from(ids));
-        let vectors = Arc::new(
-            FixedSizeListArray::try_new_from_values(Float32Array::from(values), DIM as i32)
-                .unwrap(),
-        );
-        let schema: Arc<_> = Schema::new(vec![
-            Field::new("id", DataType::UInt64, false),
-            Field::new("vector", vectors.data_type().clone(), false),
-        ])
-        .into();
-        let batch = RecordBatch::try_new(schema.clone(), vec![ids, vectors]).unwrap();
         (batch, schema)
     }
 
@@ -993,216 +960,6 @@ mod tests {
             "Expected partition split to increase partitions from 2 to 3 for {}, got stats: {}",
             description,
             final_ctx.stats_json()
-        );
-    }
-
-    async fn shrink_smallest_partition(
-        dataset: &mut Dataset,
-        index_name: &str,
-        expected_after_join: usize,
-    ) -> (usize, usize, usize) {
-        const ROWS_TO_APPEND_FOR_JOIN: usize = 32;
-        let row_count_before = dataset.count_all_rows().await.unwrap();
-        let index_ctx = load_vector_index_context(dataset, "vector", index_name).await;
-        let partitions = index_ctx.stats()["indices"][0]["partitions"]
-            .as_array()
-            .expect("partitions should be present");
-        let (partition_idx, _size) = partitions
-            .iter()
-            .enumerate()
-            .filter_map(|(idx, part)| part["size"].as_u64().map(|size| (idx, size)))
-            .filter(|(_, size)| *size > 1)
-            .min_by_key(|(_, size)| *size)
-            .expect("should have at least one partition with joinable rows");
-
-        let row_ids = load_partition_row_ids(index_ctx.ivf(), partition_idx).await;
-        assert!(
-            row_ids.len() > 1,
-            "Partition {} should have removable rows",
-            partition_idx
-        );
-
-        let rows = dataset
-            .take_rows(&row_ids, dataset.schema().clone())
-            .await
-            .unwrap();
-        let ids = rows["id"].as_primitive::<UInt64Type>().values();
-        let template_values = rows["vector"]
-            .as_fixed_size_list()
-            .value(0)
-            .as_primitive::<Float32Type>()
-            .values()
-            .to_vec();
-
-        delete_ids(dataset, &ids[1..]).await;
-        compact_after_deletions(dataset).await;
-
-        append_constant_vector(dataset, ROWS_TO_APPEND_FOR_JOIN, &template_values).await;
-        dataset
-            .optimize_indices(&OptimizeOptions::new())
-            .await
-            .unwrap();
-
-        let post_ctx = load_vector_index_context(dataset, "vector", index_name).await;
-        let post_partitions = post_ctx.num_partitions();
-        assert_eq!(
-            post_partitions,
-            expected_after_join,
-            "Expected partitions to be at most {} after join, got stats: {}",
-            expected_after_join,
-            post_ctx.stats_json()
-        );
-
-        let row_count_after = dataset.count_all_rows().await.unwrap();
-        debug_assert!(
-            row_count_before + ROWS_TO_APPEND_FOR_JOIN >= row_count_after,
-            "row count should not increase after delete + append"
-        );
-        let deleted_rows = row_count_before + ROWS_TO_APPEND_FOR_JOIN - row_count_after;
-
-        (deleted_rows, ROWS_TO_APPEND_FOR_JOIN, post_partitions)
-    }
-
-    async fn append_constant_vector(dataset: &mut Dataset, rows: usize, template: &[f32]) {
-        append_constant_vector_with_params(dataset, rows, template, None).await;
-    }
-
-    async fn append_constant_vector_with_params(
-        dataset: &mut Dataset,
-        rows: usize,
-        template: &[f32],
-        write_params: Option<WriteParams>,
-    ) {
-        assert_eq!(
-            template.len(),
-            DIM,
-            "Template vector should have {} dimensions",
-            DIM
-        );
-
-        let start_id = dataset.count_all_rows().await.unwrap() as u64;
-        let ids = Arc::new(UInt64Array::from_iter_values(
-            start_id..start_id + rows as u64,
-        ));
-        let mut appended_values = Vec::with_capacity(rows * DIM);
-        for _ in 0..rows {
-            appended_values.extend_from_slice(template);
-        }
-        let vectors = Arc::new(
-            FixedSizeListArray::try_new_from_values(
-                Float32Array::from(appended_values),
-                DIM as i32,
-            )
-            .unwrap(),
-        );
-        let schema = Arc::new(Schema::new(vec![
-            Field::new("id", DataType::UInt64, false),
-            Field::new("vector", vectors.data_type().clone(), false),
-        ]));
-        let batch = RecordBatch::try_new(schema.clone(), vec![ids, vectors]).unwrap();
-        let batches = RecordBatchIterator::new(vec![Ok(batch)], schema);
-        let params = write_params.map(|mut params| {
-            params.mode = WriteMode::Append;
-            params
-        });
-        dataset.append(batches, params).await.unwrap();
-    }
-
-    #[allow(clippy::too_many_arguments)]
-    async fn append_and_verify_append_phase(
-        dataset: &mut Dataset,
-        index_name: &str,
-        template: &[f32],
-        rows_to_append: usize,
-        expected_partitions: usize,
-        expected_total_rows: usize,
-        expected_index_count: usize,
-        expect_split: bool,
-    ) {
-        append_constant_vector(dataset, rows_to_append, template).await;
-        dataset
-            .optimize_indices(&OptimizeOptions::new())
-            .await
-            .unwrap();
-
-        let stats_json = dataset.index_statistics(index_name).await.unwrap();
-        let stats: serde_json::Value = serde_json::from_str(&stats_json).unwrap();
-
-        let indices = stats["indices"]
-            .as_array()
-            .expect("indices array should exist");
-        if expect_split {
-            assert_eq!(
-                indices.len(),
-                expected_index_count,
-                "Expected {} index entries after split, got {}, stats: {}",
-                expected_index_count,
-                indices.len(),
-                stats
-            );
-        } else {
-            assert!(
-                indices.len() >= expected_index_count,
-                "Expected at least {} index entries after append, got {}, stats: {}",
-                expected_index_count,
-                indices.len(),
-                stats
-            );
-        }
-        assert!(
-            stats["num_indices"].as_u64().unwrap() as usize >= expected_index_count,
-            "num_indices should be at least {}, stats: {}",
-            expected_index_count,
-            stats
-        );
-        assert_eq!(
-            stats["num_indexed_rows"].as_u64().unwrap() as usize,
-            expected_total_rows,
-            "Total indexed rows mismatch after append"
-        );
-
-        let base_index = indices
-            .iter()
-            .max_by_key(|entry| entry["num_partitions"].as_u64().unwrap_or(0))
-            .expect("at least one index entry should exist");
-        assert_eq!(
-            base_index["num_partitions"].as_u64().unwrap() as usize,
-            expected_partitions,
-            "Partition count mismatch after append"
-        );
-
-        if expected_index_count == 1 {
-            let partitions = base_index["partitions"]
-                .as_array()
-                .expect("partitions should exist");
-            assert_eq!(
-                partitions.len(),
-                expected_partitions,
-                "Expected {} partitions, found {}",
-                expected_partitions,
-                partitions.len()
-            );
-            let partition_sizes: Vec<usize> = partitions
-                .iter()
-                .map(|part| part["size"].as_u64().unwrap() as usize)
-                .collect();
-            let total_partition_rows: usize = partition_sizes.iter().sum();
-            assert_eq!(
-                total_partition_rows, expected_total_rows,
-                "Partition sizes should sum to total rows: {:?}",
-                partition_sizes
-            );
-        } else {
-            assert!(
-                !expect_split,
-                "Split should result in a single merged index"
-            );
-        }
-
-        assert_eq!(
-            dataset.count_all_rows().await.unwrap(),
-            expected_total_rows,
-            "Dataset row count mismatch after append"
         );
     }
 
@@ -1616,79 +1373,9 @@ mod tests {
         test_delete_all_rows(params).await;
     }
 
-    #[rstest]
-    #[case(4, DistanceType::L2, 0.9)]
-    #[case(4, DistanceType::Cosine, 0.9)]
-    #[case(4, DistanceType::Dot, 0.85)]
-    #[tokio::test]
-    async fn test_build_ivf_pq(
-        #[case] nlist: usize,
-        #[case] distance_type: DistanceType,
-        #[case] recall_requirement: f32,
-    ) {
-        let ivf_params = IvfBuildParams::new(nlist);
-        let pq_params = PQBuildParams::default();
-        let params = VectorIndexParams::with_ivf_pq_params(distance_type, ivf_params, pq_params)
-            .version(crate::index::vector::IndexFileVersion::Legacy)
-            .clone();
-        test_index(params.clone(), nlist, recall_requirement, None).await;
-        if distance_type == DistanceType::Cosine {
-            test_index_multivec(params.clone(), nlist, recall_requirement).await;
-        }
-        test_distance_range(Some(params.clone()), nlist).await;
-        // PQ performs worse on farther vectors, so if we delete the many nearest vectors, the recall will be lower
-        // lower the recall requirement in remap case for PQ, because it deletes half of the vectors
-        test_remap(params, nlist, recall_requirement * 0.9).await;
-    }
-
-    #[rstest]
-    #[case(1, DistanceType::L2, 0.9)]
-    #[case(1, DistanceType::Cosine, 0.9)]
-    #[case(1, DistanceType::Dot, 0.85)]
-    #[case(4, DistanceType::L2, 0.9)]
-    #[case(4, DistanceType::Cosine, 0.9)]
-    #[case(4, DistanceType::Dot, 0.85)]
-    #[tokio::test]
-    async fn test_build_ivf_pq_v3(
-        #[case] nlist: usize,
-        #[case] distance_type: DistanceType,
-        #[case] recall_requirement: f32,
-    ) {
-        let ivf_params = IvfBuildParams::new(nlist);
-        let pq_params = PQBuildParams::default();
-        let params = VectorIndexParams::with_ivf_pq_params(distance_type, ivf_params, pq_params);
-        test_index(params.clone(), nlist, recall_requirement, None).await;
-        if distance_type == DistanceType::Cosine {
-            test_index_multivec(params.clone(), nlist, recall_requirement).await;
-        }
-        test_distance_range(Some(params.clone()), nlist).await;
-        // PQ performs worse on farther vectors, so if we delete the many nearest vectors, the recall will be lower
-        // lower the recall requirement in remap case for PQ, because it deletes half of the vectors
-        test_remap(params.clone(), nlist, recall_requirement * 0.9).await;
-        test_delete_all_rows(params).await;
-    }
-
-    #[rstest]
-    #[case(4, DistanceType::L2, 0.85)]
-    #[case(4, DistanceType::Cosine, 0.85)]
-    #[case(4, DistanceType::Dot, 0.75)]
-    #[tokio::test]
-    async fn test_build_ivf_pq_4bit(
-        #[case] nlist: usize,
-        #[case] distance_type: DistanceType,
-        #[case] recall_requirement: f32,
-    ) {
-        let ivf_params = IvfBuildParams::new(nlist);
-        let pq_params = PQBuildParams::new(32, 4);
-        let params = VectorIndexParams::with_ivf_pq_params(distance_type, ivf_params, pq_params);
-        test_index(params.clone(), nlist, recall_requirement, None).await;
-        if distance_type == DistanceType::Cosine {
-            test_index_multivec(params.clone(), nlist, recall_requirement).await;
-        }
-        // PQ performs worse on farther vectors, so if we delete the many nearest vectors, the recall will be lower
-        // lower the recall requirement in remap case for PQ, because it deletes half of the vectors
-        test_remap(params, nlist, recall_requirement * 0.9).await;
-    }
+    // NOTE: test_build_ivf_pq, test_build_ivf_pq_v3, and test_build_ivf_pq_4bit have been
+    // moved to integration tests (tests/vector_index/ivf_pq.rs) because they are slow
+    // and benefit from running with optimization enabled.
 
     #[rstest]
     #[case(4, DistanceType::L2, 0.85)]
@@ -1785,58 +1472,9 @@ mod tests {
         test_remap(params, nlist, recall_requirement).await;
     }
 
-    #[rstest]
-    #[case(4, DistanceType::L2, 0.9)]
-    #[case(4, DistanceType::Cosine, 0.9)]
-    #[case(4, DistanceType::Dot, 0.85)]
-    #[tokio::test]
-    async fn test_create_ivf_hnsw_pq(
-        #[case] nlist: usize,
-        #[case] distance_type: DistanceType,
-        #[case] recall_requirement: f32,
-    ) {
-        let ivf_params = IvfBuildParams::new(nlist);
-        let pq_params = PQBuildParams::default();
-        let hnsw_params = HnswBuildParams::default();
-        let params = VectorIndexParams::with_ivf_hnsw_pq_params(
-            distance_type,
-            ivf_params,
-            hnsw_params,
-            pq_params,
-        );
-        test_index(params.clone(), nlist, recall_requirement, None).await;
-        if distance_type == DistanceType::Cosine {
-            test_index_multivec(params.clone(), nlist, recall_requirement).await;
-        }
-        // PQ performs worse on farther vectors, so if we delete the many nearest vectors, the recall will be lower
-        // lower the recall requirement in remap case for PQ, because it deletes half of the vectors
-        test_remap(params, nlist, recall_requirement * 0.9).await;
-    }
-
-    #[rstest]
-    #[case(4, DistanceType::L2, 0.85)]
-    #[case(4, DistanceType::Cosine, 0.85)]
-    #[case(4, DistanceType::Dot, 0.8)]
-    #[tokio::test]
-    async fn test_create_ivf_hnsw_pq_4bit(
-        #[case] nlist: usize,
-        #[case] distance_type: DistanceType,
-        #[case] recall_requirement: f32,
-    ) {
-        let ivf_params = IvfBuildParams::new(nlist);
-        let pq_params = PQBuildParams::new(32, 4);
-        let hnsw_params = HnswBuildParams::default();
-        let params = VectorIndexParams::with_ivf_hnsw_pq_params(
-            distance_type,
-            ivf_params,
-            hnsw_params,
-            pq_params,
-        );
-        test_index(params.clone(), nlist, recall_requirement, None).await;
-        if distance_type == DistanceType::Cosine {
-            test_index_multivec(params, nlist, recall_requirement).await;
-        }
-    }
+    // NOTE: test_create_ivf_hnsw_pq and test_create_ivf_hnsw_pq_4bit have been moved to
+    // integration tests (tests/vector_index/ivf_hnsw.rs) because they are slow and benefit
+    // from running with optimization enabled.
 
     async fn test_index_multivec(params: VectorIndexParams, nlist: usize, recall_requirement: f32) {
         // we introduce XTR for performance, which would reduce the recall a little bit
@@ -2428,402 +2066,11 @@ mod tests {
         assert!(pq_meta.buffer_index().is_some());
     }
 
-    #[tokio::test]
-    async fn test_optimize_with_empty_partition() {
-        let test_dir = TempStrDir::default();
-        let test_uri = test_dir.as_str();
-        let (mut dataset, _) = generate_test_dataset::<Float32Type>(test_uri, 0.0..1.0).await;
-
-        let num_rows = dataset.count_all_rows().await.unwrap();
-        let nlist = num_rows + 2;
-        let centroids = generate_random_array(nlist * DIM);
-        let ivf_centroids = FixedSizeListArray::try_new_from_values(centroids, DIM as i32).unwrap();
-        let ivf_params =
-            IvfBuildParams::try_with_centroids(nlist, Arc::new(ivf_centroids)).unwrap();
-        let params = VectorIndexParams::with_ivf_pq_params(
-            DistanceType::Cosine,
-            ivf_params,
-            PQBuildParams::default(),
-        );
-        dataset
-            .create_index(&["vector"], IndexType::Vector, None, &params, true)
-            .await
-            .unwrap();
-
-        append_dataset::<Float32Type>(&mut dataset, 1, 0.0..1.0).await;
-        dataset
-            .optimize_indices(&OptimizeOptions::new())
-            .await
-            .unwrap();
-    }
-
-    #[tokio::test]
-    async fn test_create_index_with_many_invalid_vectors() {
-        let test_dir = TempStrDir::default();
-        let test_uri = test_dir.as_str();
-
-        // we use 8192 batch size by default, so we need to generate 8192 * 3 vectors to get 3 batches
-        // generate 3 batches, and the first batch's vectors are all with NaN
-        let num_rows = 8192 * 3;
-        let mut vectors = Vec::new();
-        for i in 0..num_rows {
-            if i < 8192 {
-                vectors.extend(std::iter::repeat_n(f32::NAN, DIM));
-            } else if i < 8192 * 2 {
-                vectors.extend(std::iter::repeat_n(rand::random::<f32>(), DIM));
-            } else {
-                vectors.extend(std::iter::repeat_n(rand::random::<f32>() * 1e20, DIM));
-            }
-        }
-        let schema = Schema::new(vec![Field::new(
-            "vector",
-            DataType::FixedSizeList(
-                Arc::new(Field::new("item", DataType::Float32, true)),
-                DIM as i32,
-            ),
-            true,
-        )]);
-        let schema = Arc::new(schema);
-        let batch = RecordBatch::try_new(
-            schema.clone(),
-            vec![Arc::new(
-                FixedSizeListArray::try_new_from_values(Float32Array::from(vectors), DIM as i32)
-                    .unwrap(),
-            )],
-        )
-        .unwrap();
-        let batches = RecordBatchIterator::new(vec![batch].into_iter().map(Ok), schema);
-        let params = WriteParams {
-            mode: WriteMode::Overwrite,
-            ..Default::default()
-        };
-        let mut dataset = Dataset::write(batches, test_uri, Some(params))
-            .await
-            .unwrap();
-
-        let params = VectorIndexParams::ivf_pq(4, 8, DIM / 8, DistanceType::Dot, 50);
-
-        dataset
-            .create_index(&["vector"], IndexType::Vector, None, &params, true)
-            .await
-            .unwrap();
-    }
-
-    #[tokio::test]
-    async fn test_remap_join_on_second_delta() {
-        const INDEX_NAME: &str = "vector_idx";
-        const BASE_ROWS_PER_PARTITION: usize = 3_000;
-        const SMALL_APPEND_ROWS: usize = 64;
-        let offsets = [-50.0, 50.0];
-
-        let test_dir = TempStrDir::default();
-        let test_uri = test_dir.as_str();
-
-        let (batch, schema) = generate_clustered_batch(BASE_ROWS_PER_PARTITION, offsets);
-        let batches = RecordBatchIterator::new(vec![Ok(batch)], schema.clone());
-        let mut dataset = Dataset::write(
-            batches,
-            test_uri,
-            Some(WriteParams {
-                mode: WriteMode::Overwrite,
-                ..Default::default()
-            }),
-        )
-        .await
-        .unwrap();
-
-        let centroids = build_centroids_for_offsets(&offsets);
-        let ivf_params = IvfBuildParams::try_with_centroids(2, centroids).unwrap();
-        let params = VectorIndexParams::with_ivf_pq_params(
-            DistanceType::L2,
-            ivf_params,
-            PQBuildParams::default(),
-        );
-        dataset
-            .create_index(
-                &["vector"],
-                IndexType::Vector,
-                Some(INDEX_NAME.to_string()),
-                &params,
-                true,
-            )
-            .await
-            .unwrap();
-
-        let template_batch = dataset
-            .take_rows(&[0], dataset.schema().clone())
-            .await
-            .unwrap();
-        let template_values = template_batch["vector"]
-            .as_fixed_size_list()
-            .value(0)
-            .as_primitive::<Float32Type>()
-            .values()
-            .to_vec();
-        let mut append_params = WriteParams {
-            max_rows_per_file: 32,
-            max_rows_per_group: 32,
-            ..Default::default()
-        };
-        append_params.mode = WriteMode::Append;
-        append_constant_vector_with_params(
-            &mut dataset,
-            SMALL_APPEND_ROWS,
-            &template_values,
-            Some(append_params),
-        )
-        .await;
-
-        dataset
-            .optimize_indices(&OptimizeOptions::new())
-            .await
-            .unwrap();
-
-        let stats_before: serde_json::Value =
-            serde_json::from_str(&dataset.index_statistics(INDEX_NAME).await.unwrap()).unwrap();
-        assert_eq!(stats_before["num_indices"].as_u64().unwrap(), 2);
-        let partitions_before: Vec<usize> = stats_before["indices"]
-            .as_array()
-            .unwrap()
-            .iter()
-            .map(|idx| idx["num_partitions"].as_u64().unwrap() as usize)
-            .collect();
-        assert_eq!(partitions_before.len(), 2);
-        let base_partition_count = partitions_before
-            .iter()
-            .copied()
-            .max()
-            .expect("expected at least one partition count");
-        assert!(base_partition_count >= 2);
-        assert!(partitions_before
-            .iter()
-            .all(|count| *count == base_partition_count));
-
-        let indices_meta = dataset.load_indices_by_name(INDEX_NAME).await.unwrap();
-        assert_eq!(indices_meta.len(), 2);
-
-        compact_files(
-            &mut dataset,
-            CompactionOptions {
-                target_rows_per_fragment: 5_000,
-                ..Default::default()
-            },
-            None,
-        )
-        .await
-        .unwrap();
-
-        let mut dataset = Dataset::open(test_uri).await.unwrap();
-        let stats_after_compaction: serde_json::Value =
-            serde_json::from_str(&dataset.index_statistics(INDEX_NAME).await.unwrap()).unwrap();
-        assert_eq!(stats_after_compaction["num_indices"].as_u64().unwrap(), 2);
-        let mut partitions_after: Vec<usize> = stats_after_compaction["indices"]
-            .as_array()
-            .unwrap()
-            .iter()
-            .map(|idx| idx["num_partitions"].as_u64().unwrap() as usize)
-            .collect();
-        partitions_after.sort_unstable();
-        assert_eq!(
-            partitions_after,
-            vec![base_partition_count, base_partition_count]
-        );
-
-        const LARGE_APPEND_ROWS: usize = 40_000;
-        append_constant_vector(&mut dataset, LARGE_APPEND_ROWS, &template_values).await;
-        dataset
-            .optimize_indices(&OptimizeOptions::new())
-            .await
-            .unwrap();
-
-        let dataset = Dataset::open(test_uri).await.unwrap();
-        let stats_after_split: serde_json::Value =
-            serde_json::from_str(&dataset.index_statistics(INDEX_NAME).await.unwrap()).unwrap();
-        assert_eq!(stats_after_split["num_indices"].as_u64().unwrap(), 1);
-        let final_partition_count = stats_after_split["indices"][0]["num_partitions"]
-            .as_u64()
-            .unwrap() as usize;
-        assert_eq!(
-            final_partition_count,
-            base_partition_count + 1,
-            "expected split to increase partitions beyond {}, got {}",
-            base_partition_count,
-            final_partition_count
-        );
-    }
-
-    #[tokio::test]
-    async fn test_spfresh_join_split() {
-        // Two join cycles followed by three append cycles:
-        // 1. Each deletion shrinks the smallest partition and verifies the partition count.
-        // 2. Append #1 (10k rows) creates a delta index without splitting.
-        // 3. Append #2 and #3 (40k rows each) trigger splits, forcing merges and validating partition sizes.
-
-        const INDEX_NAME: &str = "vector_idx";
-        const NLIST: usize = 3;
-        const FIRST_APPEND_ROWS: usize = 10_000;
-        const SECOND_APPEND_ROWS: usize = 30_000;
-        const THIRD_APPEND_ROWS: usize = 35_000;
-
-        let test_dir = TempStrDir::default();
-        let test_uri = test_dir.as_str();
-
-        // Two small clusters (for joins) and two large clusters (for splits).
-        let cluster_sizes = [100, 4_000, 4_000];
-        let total_rows: usize = cluster_sizes.iter().sum();
-
-        let mut centroid_values = Vec::new();
-        for i in 0..NLIST {
-            for j in 0..DIM {
-                centroid_values.push(if j == 0 { (i as f32) * 10.0 } else { 0.0 });
-            }
-        }
-        let centroids = Arc::new(
-            FixedSizeListArray::try_new_from_values(
-                Float32Array::from(centroid_values),
-                DIM as i32,
-            )
-            .unwrap(),
-        );
-
-        let mut ids = Vec::new();
-        let mut vector_values = Vec::new();
-        let mut current_id = 0u64;
-        for (cluster_idx, &size) in cluster_sizes.iter().enumerate() {
-            let centroid_base = (cluster_idx as f32) * 10.0;
-            for _ in 0..size {
-                ids.push(current_id);
-                current_id += 1;
-                for j in 0..DIM {
-                    vector_values.push(if j == 0 {
-                        centroid_base + (current_id % 100) as f32 * 0.005
-                    } else {
-                        (current_id % 50) as f32 * 0.01
-                    });
-                }
-            }
-        }
-
-        let ids_array = Arc::new(UInt64Array::from(ids.clone()));
-        let vectors = Arc::new(
-            FixedSizeListArray::try_new_from_values(Float32Array::from(vector_values), DIM as i32)
-                .unwrap(),
-        );
-        let schema = Arc::new(Schema::new(vec![
-            Field::new("id", DataType::UInt64, false),
-            Field::new("vector", vectors.data_type().clone(), false),
-        ]));
-        let batch = RecordBatch::try_new(schema.clone(), vec![ids_array, vectors]).unwrap();
-        let batches = RecordBatchIterator::new(vec![Ok(batch)], schema);
-
-        let mut dataset = Dataset::write(
-            batches,
-            test_uri,
-            Some(WriteParams {
-                mode: crate::dataset::WriteMode::Overwrite,
-                ..Default::default()
-            }),
-        )
-        .await
-        .unwrap();
-
-        let ivf_params = IvfBuildParams::try_with_centroids(NLIST, centroids).unwrap();
-        let params = VectorIndexParams::with_ivf_pq_params(
-            DistanceType::L2,
-            ivf_params,
-            PQBuildParams::default(),
-        );
-        dataset
-            .create_index(
-                &["vector"],
-                IndexType::Vector,
-                Some(INDEX_NAME.to_string()),
-                &params,
-                true,
-            )
-            .await
-            .unwrap();
-
-        // Template vector from the first large cluster for deterministic appends.
-        let template_id = (cluster_sizes[0] + cluster_sizes[1]) as u64;
-        let template_batch = dataset
-            .take_rows(&[template_id], dataset.schema().clone())
-            .await
-            .unwrap();
-        let template_values = template_batch["vector"]
-            .as_fixed_size_list()
-            .value(0)
-            .as_primitive::<Float32Type>()
-            .values()
-            .to_vec();
-        assert_eq!(
-            template_values.len(),
-            DIM,
-            "Template vector should match DIM"
-        );
-
-        let mut expected_partitions = NLIST;
-        let mut expected_rows = total_rows;
-
-        // Two join cycles.
-        for expected_after in [NLIST - 1, NLIST - 2] {
-            let (deleted_rows, appended_rows, actual_partitions) =
-                shrink_smallest_partition(&mut dataset, INDEX_NAME, expected_after).await;
-            expected_rows = expected_rows - deleted_rows + appended_rows;
-            assert_eq!(
-                dataset.count_all_rows().await.unwrap(),
-                expected_rows,
-                "Row count mismatch after join"
-            );
-            expected_partitions = actual_partitions;
-        }
-
-        // Append #1: no split, expect a delta index.
-        let rows = FIRST_APPEND_ROWS;
-        append_and_verify_append_phase(
-            &mut dataset,
-            INDEX_NAME,
-            &template_values,
-            rows,
-            expected_partitions,
-            expected_rows + rows,
-            2,
-            false,
-        )
-        .await;
-        expected_rows += rows;
-
-        // Append #2: triggers split and merge.
-        expected_partitions += 1;
-        let rows = SECOND_APPEND_ROWS;
-        append_and_verify_append_phase(
-            &mut dataset,
-            INDEX_NAME,
-            &template_values,
-            rows,
-            expected_partitions,
-            expected_rows + rows,
-            1,
-            true,
-        )
-        .await;
-        expected_rows += rows;
-
-        // Append #3: triggers another split, remains a single merged index.
-        expected_partitions += 1;
-        let rows = THIRD_APPEND_ROWS;
-        append_and_verify_append_phase(
-            &mut dataset,
-            INDEX_NAME,
-            &template_values,
-            rows,
-            expected_partitions,
-            expected_rows + rows,
-            1,
-            true,
-        )
-        .await;
-    }
+    // NOTE: test_optimize_with_empty_partition, test_create_index_with_many_invalid_vectors,
+    // and test_remap_join_on_second_delta have been moved to integration tests
+    // (tests/vector_index/lifecycle.rs and tests/vector_index/ivf_pq.rs) because they are
+    // slow and benefit from running with optimization enabled.
+    // test_spfresh_join_split was NOT moved because it requires private API access.
 
     #[tokio::test]
     async fn test_partition_split_on_append_multivec() {
